move reducer to seperate file
#146
