+++ projects/ngx-d3timeline/src/lib/options.service.ts
@@ -5,11 +5,6 @@ import { AxisOrientations } from './axis-orientations';
 
 @Injectable({ providedIn: 'root' })
 export class OptionsService {
-  setAxisOrientations(timeOrientation: Orientation): AxisOrientations {
-    const resourceOrientation = this.flipOrientation(timeOrientation);
-    return { time: timeOrientation, resource: resourceOrientation };
-  }
-
   getTranslation(
     range: number,
     orientation: Orientation,
@@ -18,12 +13,6 @@ export class OptionsService {
     return orientation === Orientation.Vertical
       ? `translate(${timelineView.left}, ${range})`
       : `translate(${range}, ${timelineView.top})`;
-  }
-
-  private flipOrientation(orientation: Orientation) {
-    return orientation === Orientation.Vertical
-      ? Orientation.Horizontal
-      : Orientation.Vertical;
   }
 
   getTickLineEnd(lineOffset: number, orientation: Orientation) {
