+++ projects/ngx-d3timeline/src/lib/utils.ts
@@ -1,0 +1,15 @@
+import { Orientation } from './orientation';
+import { AxisOrientations } from './axis-orientations';
+
+export function setAxisOrientations(
+  timeOrientation: Orientation
+): AxisOrientations {
+  const resourceOrientation = this.flipOrientation(timeOrientation);
+  return { time: timeOrientation, resource: resourceOrientation };
+}
+
+export function flipOrientation(orientation: Orientation) {
+  return orientation === Orientation.Vertical
+    ? Orientation.Horizontal
+    : Orientation.Vertical;
+}
